[AS7-3271] Make all STORAGE attributes read-only for a domain mode server
